client: move tests to distinct module
Moves the `janus_client::tests` module into a separate file.

As we have found elsewhere, putting tests into separate files improves
compilaton speed ([citation needed]), and makes it easier to navigate
sources.
